Incogine Development for Windows Support
This will make developers have ease while working on Incogine with a Windows machine, for real this time because it works on my end.
I got used into my Macbook which is in macOS so virtualizing with Windows in it really helps and made me look into this project to make sure it can be developed, tested and run anywhere.
Next forward, we are looking forward to make this project works on Linux but of course the first will be on Ubuntu.
Sooner or later in every Linux distros we can find, of course Android duh!
Looking forward on Incogine's development after every project being rewritten like many times on doing a game engine, this will be my final game engine attempt but hey, it's gonna be worth it.